refactor class name and location
