+++ app/src/main/java/za/co/app/budgetbee/base/AppCompatBaseActivity.kt
@@ -1,13 +1,12 @@
 package za.co.app.budgetbee.base
 
 import android.os.Bundle
-import android.os.PersistableBundle
 import androidx.appcompat.app.AppCompatActivity
 import za.co.app.budgetbee.BudgetBeeApplication
 
 abstract class AppCompatBaseActivity : AppCompatActivity() {
-    override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
-        super.onCreate(savedInstanceState, persistentState)
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
         BudgetBeeApplication.instance.feather.injectFields(this)
     }
 }
